@@ -22,12 +22,12 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
-        "triagens/arangodb": "dev-master"
+        "php": ">=5.5",
+        "triagens/arangodb": "~2.8.1"
     },
     "require-dev": {
-        "silex/silex": "~1.2.0",
-        "phpunit/phpunit": "4.5.*"
+        "silex/silex": "~1.3.5",
+        "phpunit/phpunit": "5.3.*"
     },
     "autoload": {
         "psr-0": {
